Factorize detection merge
