+++ examples/anthropogenic-activities/merge_detections.py
@@ -9,7 +9,7 @@ import numpy as np
 import pandas as pd
 
 sys.path.insert(1, '../..')
-from helpers.functions_for_examples import get_categories
+from helpers.functions_for_examples import get_categories, merge_adjacent_detections, read_dets_and_aoi
 import helpers.misc as misc
 
 from loguru import logger
@@ -63,27 +63,7 @@ if __name__ == "__main__":
         logger.success(f"Done! All files already exist in folder {OUTPUT_DIR}. Exiting.")
         sys.exit(0)
 
-    logger.info("Loading split AoI tiles as a GeoPandas DataFrame...")
-    tiles_gdf = gpd.read_file('split_aoi_tiles.geojson')
-    tiles_gdf = tiles_gdf.to_crs(2056)
-    if 'year_tile' in tiles_gdf.keys(): 
-        tiles_gdf['year_tile'] = tiles_gdf.year_tile.astype(int)
-    logger.success(f"Done! {len(tiles_gdf)} features were found.")
-
-    logger.info("Loading detections as a GeoPandas DataFrame...")
-
-    detections_gdf = gpd.GeoDataFrame()
-
-    for dataset, dets_file in DETECTION_FILES.items():
-        detections_ds_gdf = gpd.read_file(dets_file)    
-        detections_ds_gdf[f'dataset'] = dataset
-        detections_gdf = pd.concat([detections_gdf, detections_ds_gdf], axis=0, ignore_index=True)
-    detections_gdf = detections_gdf.to_crs(2056)
-    detections_gdf['area'] = detections_gdf.area 
-    detections_gdf['det_id'] = detections_gdf.index
-    if 'year_det' in detections_gdf.keys(): 
-        detections_gdf['year_det'] = detections_gdf.year_det.astype(int)
-    logger.success(f"Done! {len(detections_gdf)} features were found.")
+    tiles_gdf, detections_gdf = read_dets_and_aoi(DETECTION_FILES)
 
     # Merge features
     logger.info(f"Merge adjacent polygons overlapping tiles with a buffer of {DISTANCE} m...")
@@ -91,58 +71,8 @@ if __name__ == "__main__":
 
     # Process detection by year
     for year in detections_gdf.year_det.unique():
-        detections_by_year_gdf = detections_gdf[detections_gdf['year_det']==year]
-
-        detections_buffer_gdf = detections_by_year_gdf.copy()
-        detections_buffer_gdf['geometry'] = detections_by_year_gdf.geometry.buffer(DISTANCE, resolution=2)
-
-        # Saves the id of polygons contained entirely within the tile (no merging with adjacent tiles), to avoid merging them if they are at a distance of less than thd  
-        detections_tiles_join_gdf = gpd.sjoin(tiles_gdf, detections_buffer_gdf, how='left', predicate='contains')
-        remove_det_list = detections_tiles_join_gdf.det_id.unique().tolist()
-
-        detections_overlap_tiles_gdf = detections_by_year_gdf[~detections_by_year_gdf.det_id.isin(remove_det_list)].drop_duplicates(subset=['det_id'], ignore_index=True)
-        detections_within_tiles_gdf = detections_by_year_gdf[detections_by_year_gdf.det_id.isin(remove_det_list)].drop_duplicates(subset=['det_id'], ignore_index=True)
-
-        # Merge polygons within the thd distance
-        detections_overlap_tiles_gdf.loc[:, 'geometry'] = detections_overlap_tiles_gdf.buffer(DISTANCE, resolution=2)
-        detections_dissolve_gdf = detections_overlap_tiles_gdf[['det_id', 'geometry']].dissolve(as_index=False)
-        detections_merge_gdf = detections_dissolve_gdf.explode(ignore_index=True)
-        del detections_dissolve_gdf, detections_overlap_tiles_gdf
-
-        if detections_merge_gdf.isnull().values.any():
-            detections_merge_gdf = gpd.GeoDataFrame()
-        else:
-            detections_merge_gdf.geometry = detections_merge_gdf.buffer(-DISTANCE, resolution=2)
-    
-        # Spatially join merged detection with raw ones to retrieve relevant information (score, area,...)
-        detections_merge_gdf['index_merge'] = detections_merge_gdf.index
-        detections_join_gdf = gpd.sjoin(detections_merge_gdf, detections_by_year_gdf, how='inner', predicate='intersects')
-
-        det_class_all = []
-        det_score_all = []
-
-        for id in detections_merge_gdf.index_merge.unique():
-            detections_by_year_gdf = detections_join_gdf.copy()
-            detections_by_year_gdf = detections_by_year_gdf[(detections_by_year_gdf['index_merge']==id)]
-            detections_by_year_gdf.rename(columns={'score_left': 'score'}, inplace=True)
-            det_score_all.append(detections_by_year_gdf['score'].mean())
-            detections_by_year_gdf = detections_by_year_gdf.dissolve(by='det_class', aggfunc='sum', as_index=False)
-            # Keep class of largest det
-            if len(detections_by_year_gdf) > 0:
-                detections_by_year_gdf['det_class'] = detections_by_year_gdf.loc[detections_by_year_gdf['area'] == detections_by_year_gdf['area'].max(), 
-                                                                'det_class'].iloc[0]    
-                det_class = detections_by_year_gdf['det_class'].drop_duplicates().tolist()
-            else:
-                det_class = [0]
-            det_class_all.append(det_class[0])
-
-        detections_merge_gdf['det_class'] = det_class_all
-        detections_merge_gdf['score'] = det_score_all
-
-        complete_merge_dets_gdf = pd.merge(detections_merge_gdf, detections_join_gdf[['index_merge', 'year_det'] + ([] if 'dataset' in detections_merge_gdf.columns else ['dataset'])], on='index_merge')
+        complete_merge_dets_gdf, detections_within_tiles_gdf = merge_adjacent_detections(detections_gdf, tiles_gdf, year, DISTANCE)
         detections_all_years_gdf = pd.concat([detections_all_years_gdf, complete_merge_dets_gdf, detections_within_tiles_gdf], ignore_index=True)
-
-        del complete_merge_dets_gdf, detections_merge_gdf, detections_by_year_gdf, detections_within_tiles_gdf, detections_join_gdf
 
     # get classe ids
     CATEGORIES = os.path.join('category_ids.json')
